requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.104.1
frozenlist==1.4.0
h11==0.14.0
idna==3.4
multidict==6.0.4
openai==0.27.9
outcome==1.2.0
pydantic==2.4.2
pydantic_core==2.10.1
PySocks==1.7.1
python-dotenv==1.0.0
regex==2023.8.8
requests==2.31.0
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
starlette==0.27.0
tiktoken==0.4.0
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.8.0
urllib3==2.0.4
uvicorn==0.24.0.post1
wsproto==1.2.0
yarl==1.9.2
